+++ diag/mfg/mfg_kpt_test.py
@@ -24,138 +24,6 @@ from libmtp_ctrl import mtp_ctrl
 from libpro_srv_db import pro_srv_db
 
 
-def get_mtp_logfile(mtp_mgmt_ctrl, log_dir, mtp_id, mtp_test_summary):
-    mtp_mgmt_cfg = mtp_mgmt_ctrl.get_mgmt_cfg()
-    ipaddr = mtp_mgmt_cfg[0]
-    userid = mtp_mgmt_cfg[1]
-    passwd = mtp_mgmt_cfg[2]
-
-    mtp_mgmt_ctrl.cli_log_inf("Collecting log files started", level=0)
-
-    # create the log subdir
-    log_timestamp = libmfg_utils.get_timestamp()
-    sub_dir = MTP_DIAG_Logfile.MFG_KPT_LOG_DIR.format(mtp_id, log_timestamp)
-    mtp_mgmt_ctrl.mtp_mgmt_exec_cmd(MFG_DIAG_CMDS.MFG_MK_DIR_FMT.format(log_dir+sub_dir))
-
-    # log pkg filename
-    log_pkg_file = log_dir + MTP_DIAG_Logfile.MFG_KPT_LOG_PKG_FILE.format(mtp_id, log_timestamp)
-
-    # need to be sync'd with cleanup.sh
-    test_onboard_log_files = MTP_DIAG_Logfile.ONBOARD_KPT_LOG_FILES
-
-    # kpt test logs
-    logfile_list = list()
-    # diag onboard log files
-    cmd = "mv {:s} {:s}".format(test_onboard_log_files, log_dir+sub_dir)
-    mtp_mgmt_ctrl.mtp_mgmt_exec_cmd(cmd)
-
-    # all the test logs
-    test_log_file = "{:s}test_kpt.log".format(log_dir+sub_dir)
-
-    # pkg the onboard logs
-    cmd = MFG_DIAG_CMDS.MFG_LOG_PKG_FMT.format(log_pkg_file, log_dir, sub_dir)
-    mtp_mgmt_ctrl.mtp_mgmt_exec_cmd(cmd)
-
-    local_test_log_file = "log/{:s}_test_kpt.log".format(mtp_id)
-    libmfg_utils.network_get_file(ipaddr, userid, passwd, local_test_log_file, test_log_file)
-
-    with open(local_test_log_file, 'r') as fp:
-        buf = fp.read()
-    nic_fail_reg_exp = MTP_DIAG_Report.NIC_DIAG_REGRESSION_RSLT_RE.format(MTP_DIAG_Report.NIC_DIAG_REGRESSION_FAIL)
-    nic_pass_reg_exp = MTP_DIAG_Report.NIC_DIAG_REGRESSION_RSLT_RE.format(MTP_DIAG_Report.NIC_DIAG_REGRESSION_PASS)
-    fail_match = re.findall(nic_fail_reg_exp, buf)
-    pass_match = re.findall(nic_pass_reg_exp, buf)
-
-    for slot, nic_type, sn in fail_match + pass_match:
-        if GLB_CFG_MFG_TEST_MODE:
-            mfg_log_dir = MTP_DIAG_Logfile.DIAG_MFG_KPT_LOG_DIR_FMT.format(nic_type, sn)
-        else:
-            mfg_log_dir = MTP_DIAG_Logfile.DIAG_MFG_MODEL_KPT_LOG_DIR_FMT.format(nic_type, sn)
-
-        cmd = MFG_DIAG_CMDS.MFG_MK_DIR_FMT.format(mfg_log_dir)
-        os.system(cmd)
-        # copy the onboard logs
-        ts = libmfg_utils.get_timestamp()
-        qa_log_pkg_file = mfg_log_dir + os.path.basename(log_pkg_file)
-        mtp_mgmt_ctrl.cli_log_inf("Collecting {:s} log files {:s}".format(sn, qa_log_pkg_file))
-        libmfg_utils.network_get_file(ipaddr, userid, passwd, qa_log_pkg_file, log_pkg_file)
-
-    for slot, nic_type, sn in fail_match:
-        mtp_test_summary.append((slot, sn, nic_type, False))
-
-    for slot, nic_type, sn in pass_match:
-        mtp_test_summary.append((slot, sn, nic_type, True))
-
-    # clear the onboard logs
-    logfile_list.append(log_pkg_file)
-    logfile_list.append(log_dir+sub_dir)
-    cmd = "rm -rf {:s}".format(" ".join(logfile_list))
-    mtp_mgmt_ctrl.mtp_mgmt_exec_cmd(cmd)
-    mtp_mgmt_ctrl.cli_log_inf("Collecting log files complete", level=0)
-
-    return local_test_log_file
-
-
-def mfg_report(mtp_id, mtp_start_ts, mtp_stop_ts, test_log_file):
-    mtp_cli_id_str = libmfg_utils.id_str(mtp = mtp_id)
-    duration = mtp_stop_ts - mtp_start_ts
-
-    with open(test_log_file, 'r') as fp:
-        buf = fp.read()
-
-    # MTP related error, don't post any report
-    if MTP_DIAG_Report.MTP_DIAG_REGRESSION_FAIL in buf:
-        libmfg_utils.cli_inf(mtp_cli_id_str + "MTP Setup fails, no report will be generated")
-        return
-
-    libmfg_utils.cli_inf(mtp_cli_id_str + "Start posting test report")
-    if MTP_DIAG_Report.NIC_DIAG_REGRESSION_FAIL in buf:
-        nic_fail_reg_exp = MTP_DIAG_Report.NIC_DIAG_REGRESSION_RSLT_RE.format(MTP_DIAG_Report.NIC_DIAG_REGRESSION_FAIL)
-        match = re.findall(nic_fail_reg_exp, buf)
-        for slot, nic_type, sn in match:
-            test_list = list()
-            test_rslt_list = list()
-            err_dsc_list = list()
-            err_code_list = list()
-            nic_cli_id_str = libmfg_utils.id_str(mtp=mtp_id, nic=int(slot), base=0)
-            # find all test status
-            nic_test_rslt_reg_exp = MTP_DIAG_Report.NIC_DIAG_TEST_RSLT_RE.format(slot, sn)
-            sub_match = re.findall(nic_test_rslt_reg_exp, buf)
-            for dsp, test, result in sub_match:
-                test_list.append("{:s}-{:s}".format(dsp, test))
-                test_rslt_list.append(result)
-                err_dsc_list.append(nic_cli_id_str)
-                err_code_list.append(result)
-            ret = libmfg_utils.flx_web_srv_post_uut_report("KPT", nic_type, sn, "FAIL", mtp_start_ts, mtp_stop_ts, duration, test_list, test_rslt_list, err_dsc_list, err_code_list)
-            if not ret:
-                libmfg_utils.cli_inf(mtp_cli_id_str + "Post [{:s}] result to webserver failed".format(sn))
-            else:
-                libmfg_utils.cli_inf(mtp_cli_id_str + "Post [{:s}] result to webserver complete".format(sn))
-
-    if MTP_DIAG_Report.NIC_DIAG_REGRESSION_PASS in buf:
-        nic_pass_reg_exp = MTP_DIAG_Report.NIC_DIAG_REGRESSION_RSLT_RE.format(MTP_DIAG_Report.NIC_DIAG_REGRESSION_PASS)
-        match = re.findall(nic_pass_reg_exp, buf)
-        for slot, nic_type, sn in match:
-            test_list = list()
-            test_rslt_list = list()
-            err_dsc_list = list()
-            err_code_list = list()
-            nic_cli_id_str = libmfg_utils.id_str(mtp=mtp_id, nic=int(slot), base=0)
-            # find all test status
-            nic_test_rslt_reg_exp = MTP_DIAG_Report.NIC_DIAG_TEST_RSLT_RE.format(slot, sn)
-            sub_match = re.findall(nic_test_rslt_reg_exp, buf)
-            for dsp, test, result in sub_match:
-                test_list.append("{:s}-{:s}".format(dsp, test))
-                test_rslt_list.append(result)
-                err_dsc_list.append(nic_cli_id_str)
-                err_code_list.append(result)
-            ret = libmfg_utils.flx_web_srv_post_uut_report("KPT", nic_type, sn, "PASS", mtp_start_ts, mtp_stop_ts, duration, test_list, test_rslt_list, err_dsc_list, err_code_list)
-            if not ret:
-                libmfg_utils.cli_inf(mtp_cli_id_str + "Post [{:s}] result to webserver failed".format(sn))
-            else:
-                libmfg_utils.cli_inf(mtp_cli_id_str + "Post [{:s}] result to webserver complete".format(sn))
-
-
 def load_mtp_cfg():
     mtp_chassis_cfg_file_list = list()
     if not GLB_CFG_MFG_TEST_MODE:
@@ -194,9 +62,13 @@ def single_mtp_kpt_test(mtp_kpt_script_dir, mtp_mgmt_ctrl, mtp_id, mtp_test_summ
     mtp_mgmt_ctrl.cli_log_inf("MTP KPT Test complete", level=0)
     mtp_mgmt_ctrl.cli_log_inf("MTP KPT Test Duration:{:s}".format(mtp_stop_ts-mtp_start_ts), level=0)
 
-    test_log_file = get_mtp_logfile(mtp_mgmt_ctrl, mtp_kpt_script_dir, mtp_id, mtp_test_summary)
+    test_log_file = libmfg_utils.get_mtp_logfile(mtp_mgmt_ctrl, mtp_dl_script_dir, mtp_id, mtp_test_summary, FF_Stage.FF_KPT)
+    if not test_log_file:
+        mtp_mgmt_ctrl.cli_log_err("MTP Collect KPT Test result failed", level=0)
+        return
+
     if GLB_CFG_MFG_TEST_MODE:
-        mfg_report(mtp_id, mtp_start_ts, mtp_stop_ts, test_log_file)
+        libmfg_utils.mfg_report(mtp_id, mtp_start_ts, mtp_stop_ts, test_log_file, FF_Stage.FF_KPT)
 
     cmd = "rm -rf {:s}".format(test_log_file)
     os.system(cmd)
